backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==10.2.0
torch==2.2.0
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.2.1
pyyaml==6.0.1
python-dotenv==1.0.1
